Further refactoring related to separator suppression.
Status:
All basic tests run. See note below about "threePass" tests.

Of application schemas I can test, army-drrs doesn't work reasons unclear but appear to be unrelated to this change set. I have not tested vmf, as it's build file must be updated to work with the newer sbt and scala 2.12.

USMTF had to be changed so that erroneous use
of separatorSuppressionPolicy trailingEmptyStrict in it is changed to
anyEmpty.

Everything else in DFDLSchemas and the fouo schemas works.

Changes:

Got rid of everything about stop values and other unimplemented cruft.

Remove ability for any schema component to walk backrefs to a Term.

Only terms are able to "generate code" i.e., have gram (grammar)
objects. Code that resides on other schema component classes that is
needed to generate code, if that code needs properties to be resolved,
then it has to go on a subclass of Term.

Only Terms are able to resolve properties - schema components can get
access to local properties for certain warning checks, but they can't
lookup properties for resolution.

Added HasTermCheck mixin to eliminate backpointers from some places.

Some checks were being missed due to the refactoring of
GlobalElementDecl where it is no longer an ElementBase and hence, not a
Term.

Added this HasTermCheck feature by which a schema component can express
consistency checks that need the concrete Term.

Avoids need for the schema component to have a backpointer to the Term.

Added ChoiceBranchImpliedSequence

Used when a choice branch is non-scalar.
Required refactoring to make this a SeqTermBase, so that everything
about unseparated sequences applies to it.

Removed unimplemented UnorderedSequence DSOM class.

This class, as a DSOM object, is a bad idea, and when unordered
sequences are implemented, it will not want to use a UnorderedSequence
DSOM object, rather it should use an UnorderedSequence grammar
combinator, created from the SepTermBase DSOM class, based on the
dfdl:separatorKind property 'unordered'.

There is no more occursBounds nor stack for it
in the PState/UState, so the corresponding
CLI debug commands to check it have been removed.

Added threePass feature to TDMLRunner. Needed for testing some separator
scenarios.

Some tests changed to roundTrip="threePass" to get them to work.

Note that "threePass" can mask errors very easily. It needs to be avoided whenever possible.

In some cases the "threePass" tests are questionable in terms of
whether Daffodil is doing the right thing vis a vis the DFDL specification.

But they're easily identified by searching for roundTrip="threePass" in
the test modules. Such time as DFDL spec. clarifications are available
for some of these they should be revisited.

Fixed eclipse class paths for scala 2.12 IDE.

DAFFODIL-1919