Remove ability for any schema component to walk backrefs to a Term.
Only terms are able to "generate code" i.e., have gram (grammar)
objects. Code that resides on other schema component classes that is
needed to generate code, if that code needs properties to be resolved,
then it has to go on a subclass of Term.

Only terms are able to resolve properties - schema components can get
access to local properties for certain warning checks, but they can't
lookup properties for resolution.