Substantial refactoring to support sequence separators properly.
Also Merged with TDMLRunner explicit roundTrip="twoPass" feature.

Added test case to demonstrate ATO bug

OrderedSeparatedSequence
OrderedUnseparatedSequence
+ someday eventually Unordered versions of both the above.

These combinators each create a Parser and an Unparser, so:

OrderedSeparatedSequenceParser
OrderedSeparatedSequenceUnparser
OrderedUnseparatedSequenceParser
OrderedUnseparatedSequenceUnparser

The OrderedUnseparated flavors have no change to their function.

The OrderedSeparatedParser flavors are quite different in behavior.

Both depend heavily on SequenceChild and RepeatingSequenceChild classes which
carry the point-of-use information. (Should improve sharing/speed at some point)

Renamings and simplifications of grammar.

Removed the "|" operator entirely. Any place we're actually doing
alternation we should be doing so in some structured combinator.

Optimizes out NadaParser and unparser entirely now.

Invariant that NadaParser is always optimized out re-established.

Got rid of everything about stop values and other unimplemented cruft.

Remove ability for any schema component to walk backrefs to a Term.

Only terms are able to "generate code" i.e., have gram (grammar)
objects. Code that resides on other schema component classes that is
needed to generate code, if that code needs properties to be resolved,
then it has to go on a subclass of Term.

Only Terms are able to resolve properties - schema components can get
access to local properties for certain warning checks, but they can't
lookup properties for resolution.

Added HasTermCheck mixin to eliminate backpointers from some places.

Some checks were being missed due to the refactoring of
GlobalElementDecl where it is no longer an ElementBase and hence, not a
Term.

Added this HasTermCheck feature by which a schema component can express
consistency checks that need the concrete Term.

Avoids need for the schema component to have a backpointer to the Term.

Added ChoiceBranchImpliedSequence

Used when a choice branch is non-scalar.
Required refactoring to make this a SeqTermBase, so that everything
about unseparated sequences applies to it.

Removed unimplemented UnorderedSequence DSOM class.

This class, as a DSOM object, is a bad idea, and when unordered
sequences are implemented, it will not want to use a UnorderedSequence
DSOM object, rather it should use an UnorderedSequence grammar
combinator, created from the SepTermBase DSOM class, based on the
dfdl:separatorKind property 'unordered'.

There is no more occursBounds nor stack for it
in the PState/UState, so the corresponding
CLI debug commands to check it have been removed.

Added threePass feature to TDMLRunner. Needed for testing some separator
scenarios.

Some tests changed to roundTrip="threePass" to get them to work.

Note that "threePass" can mask errors very easily. It needs to be avoided whenever possible.

In some cases the "threePass" tests are questionable in terms of
whether Daffodil is doing the right thing vis a vis the DFDL specification.

But they're easily identified by searching for roundTrip="threePass" in
the test modules. Such time as DFDL spec. clarifications are available
for some of these they should be revisited.

Fixed eclipse class paths for scala 2.12 IDE.

Added tests and descriptions related to DAFFODIL-1975 change.
The fix for DAFFODIL-1975 is not done here, but tests that verify the
behavior change are in scala-debug.

NadaUnparser now throws error if it survives until runtime.
This insures all grammar terms optimize it out.

Added doc about Gram and parser/unparser methods that must implement the
optimizing out of NadaParser and NadaUnparser.

DAFFODIL-1919, DAFFODIL-1961, DAFFODIL-1975