Allow same character for escapeBlockEnd and escapeEscapeCharacter
This commit fixes the issue where the escapeBlockEnd and
escapeEscapeCharacter are the same character. This is apparently a
common thing in how Excel escapes things in CSV.

This is not a complete fix as there are still issues in cases like
escapeBlockEnd="++" and escapeEscapeCharacter='+', but there does not
seem to be an easy way to deal with these extreme edge cases cleanly in
the existing code. This issue has been documented in DAFFODIL-1972

DAFFODIL-1923