Update sbt and sbt-native-packager plugin to make jar hashes consistent
It looks like some combination of old versions of sbt and the sbt native
packager plugin caused generated jars in the zip/tar/rpms to have
different hashes. The only difference appeared to be the timestamp of
the file, but this makes it difficult to ensure that all files are the
same. Updating sbt and the sbt-native-packager plugin to the latest
version seems to correct this issue. So all jars, no matter if in a zip,
tar, rpm, or maven, should all have the exact same hash.

DAFFODIL-1938