Update Runtime to only use Vector.
Eliminate use of List and Sequence (mostly) in the runtime.

Should become a code-review check now, when reviewing runtime code, that
it use Vector as the standard immutable collection, vs. Sequence or
List.

This insures constant time access to the contents of the collection,
when indexed. The runtime mostly uses while-loops and indexing for speed
reasons (vs. map/foreach or other more scala-ish idioms).

DAFFODIL-1978