Ensure that buckets with no references are released
When we release a byte, we decreased the refCount of the bucket that
byte was in and then check to see if that buckets reference count was
zero and if it was the oldest bucket and, if so, we release buckets.

However, if the oldest bucket never had any locked bytes, then we'll
never release a byte in that bucket and bucketIndex will never equal
oldestBucketIndex. This means we'll never release the oldest bucket and
no buckets end up getting released.

Instead, just check to see if the oldest bucket has a refCount of zero
and then release buckets.

DAFFODIL-1987