Use --binary flag when generating release checksums
The sha*sum programs default to --text mode when generating checksums.
On GNU systems this isn't a problem since --text and --binary are the
same. But that isn't the case on Windows systems, which can result in
failed checksum validations. This change specifies the --binary option
when generating checksums to ensure compatability with Windows and
similar systems.

DAFFODIL-2009