Check for available data before trying to parse binarySeconds or binaryMilliseconds
If we do not check for available data, a NotEnoughData exception is
thrown when trying to get the binary long, which ends up leaking.
Instead, we should just check for enough data first.

Also, when exceptions are accidentally leaked, ensure that we discard
marks before throwing an assertion about the leak. Otherwise the mark is
never returned and we get an error about leaked marks instead of the
actual root issue of a leaked assertion.

DAFFODIL-2010