Ignore minLength when lengthKind=explicit
According to DFDL Erratum 5.18 if dfdl:lengthKind='explicit' the facet
minLength is not used to determin the target length.

Although we are ignoring minLength, it is a schema definition wraning if
the explicit length specified is out of range of minLength and it is an
error if it is out of range of textOutputMinLength.

Also fixed a small error in the range check warning.

DFDL-1989