Added max/minValidYear tunable
Due to differences in the DFDL spec and ICU4J SimpleDateFormat, we must
have SimpleDateFormat parse in lenient mode which allows the year value
to overflow with very large years into possibly negative years.

To combat this overflowing of years the max/minValidYear tunable has
been added with default values of "9999" and "0" respectively.

DAFFODIL-1946