Enabled a few Calendar tests that now work
There were a few Calendar tests that had been in debug for a while that
now seem to work (or only needed a few small fixes to the tests
themselves to work). They had been moved to debug as part of
DAFFODIL-1945 as part of an issue that a user was having.

DAFFODIL-1945