Missing terminator was reporting itself as a missing initiator.
I.e., message like Parse Error: Initiator ';' not found.
Except ";" is the terminator of the element, not the initiator.

Due to object-construction-timing stuff I think. That or scala bugs.
But sub-class overrides were not happening to provide the member from
which this name was derived. Why it got the wrong one, I don't know, I
would have expected "unbound/undefined" or something.

Just passing the parameter to constructor rather than overriding a
member to provide it, solves the problem.

DAFFODIL-2204