Add tunables for textBidi and floating properties
Many schemas may not have these rather uncommon properties defined, but
they are technically required and according to the DFDL spec missing
properties should not be defaulted. As a stop gap solution tunables have
been added for these properties that when enabled will cause an SDE if
the properties are not defined. If they are not enabled an SDW will be
shown about the missing property.

DAFFODIL-1720