Make the JAPI/SAPI DataProcessors serializable
Some tools, like Apache Spark, need the ability to serialize the
DataProcessor to be sent to cluster nodes where it is then deserialized
and used to process data. We have the save/reload methods, but Spark
does not now how to use these. So this modifies the JAPI/SAPI
DataProcessors so that they are serializable, and modifies the JAPI/SAPI
tests to serialize/deserialize each DataProcessor before being used.

DAFFODIL-1440