Avoid CircularDefinition error for appinfo
When we were detecting that an appinfo annotation did not have a source
attribute we would try to show a schema definition warning. However, the
lack of source attribute was somehow causing a CircularDefinition error
within the OOLAG/LV code when trying to print out the position of the
sequence where the error was occuring.

Instead of relying on the OOLAG/LV code to sort of reverse lookup the
sequences position, the position is now passed in as an argument to
ModelGroup.

DAFFODIL-2046

fixup

Fixed failing CLI test