Change expression result type coercion to be a warning
Commit a2f56a3f44 modified expression coercion to be strict about the
result type of an expression matching the expected type. This proved to
break a lot of schemas that were too lax, and so isn't good for
backwards compatibility. Instead, perform the coercion as we used to,
but display a warning to the user instead of an SDE if coercion is
needed. Additionally, add a tunable to change the warning to an SDE so
we can deprecate the behavior and easily change it in the future.

DAFFODIL-2054