Keep line/column info for Daffodil tests
We had been adding line and column attributes to the schemas we were
testing against in order to provide more helpful error messages. These
attributes were causing issues with the IBM DFDL processor and were
removed.

This commit allows the TDMLRunner to detect if it is running the IBM
DFDL implementation and remove these attributes as needed, while keeping
them for the Daffodil implementation.

DAFFODIL-2050