Make XML comparisions more memory efficient
Commit a3e15e665f changed the XML comparison to always use the
computeDiffOne function, rather than only when differences were found in
an XML.toString comparison. However, the computeDiffOne function is
implemented recursively, but in a way that every sibling creates a new
recursive function call stack. This meant that for very large infosets
with lots of siblings, we could hit a stack overflow when comparing
actual and expected.

This refactors that function so that it only calls itself on children,
and iterates over the children. This limits the function call stack
depth to the same as the maximum child depth, which avoids the stack
overflow.

This is functionally the same, just with better stack usage.

DAFFODIL-2059