Remove JUnit4 compile dependency
We currently have a JUnit4 compile dependency because the TDML Runner
throws a JUnit-based exception to indicate that a test should be skipped
when the current implementation does not support that test. The problem
with this is 1) we now have a compile dependency that is traditionally
only for tests, which is odd, and 2) JUnit is now distributed in our
binary helpers, and so requires licensing changes.

Neither of these are ideal, so this patch removes that compile
dependency. It does so by using reflection to create the appropriate
JUnit "skip this test" exception if JUnit is on the classpath (i.e.
we're probably running a JUnit test). If JUnit is not on the classpath,
then we just throws a standard TDML exception and lets the caller handle
it however it makes sense for them.

DAFFODIL-2067