Remove extraneous XML loads for repeating include/imports
We already have logic to detect when the same file has been
included/imported so that we can prevent circular imports. However, we
defined requiredEvaluations on the iiSchemaFile, which forces a load and
parse of the schema regardless if it had already been loaded or not. For
large schemas with lots of imports of the same thing (e.g. multiple
schema files importing a common schema), this could result in a lot of
duplicated effort.

Instead, define the requiredEvaluations on the iiSchemaFileMaybe, which
will only load the schema if it hasn't been seen.

On some large schemas with lots of repeating imports, this saw a
reduction in schema compile time of about 15%.

DAFFODIL-2071