Ensure content/valueLength expressions are not considered constants
Although it should result in an error, calling contentLength() or
valueLength() on an immediate parent does not always trigger the normal
code paths that determine that an expression is not a constant. This can
lead to an unknown state and ultimately a NullPointerException. By
overriding the contentLength and valueLength methods in the FakeDINode,
we ensure that if this special case occurs we correctly determine this
expression is not a constant.

Also removes debugging information from the error message that results
when this expression is eventually evaluated.

DAFFODIL-2078