Do not abort when an import/include fails
When using XML Catalogs, it's very easy for import/includes to
resolve to URI's that do not exist. This results in an IOException
when trying to open the stream. There are also other cases related
to IO that could causing opening of the stream to fail. When these
failures happen, we currently abort. This is wrong, since these
exception are expected to be able to happen.

Instead of aborting with an assert when an exception occurs, we should
instead just return null. Xerces then figures out what the problem is
and returns an error that Daffodil handles and converts to and SDE.

DAFFODIL-2126