Switch XML/JSON infoset inputters/outputters to use streams
The API for XML and JSON Infoset Inputters/Outputters requires that
users pass in a Reader or a Writer. The problem is that Readers/Writers
must be assigned an encoding, which the user may not always know. This
can result in situations where XML infosets are read in a different
encoding than how the XML was actually encoded, and can even lead to
cases where the XML preamble is completely ignored. Making things worse,
we don't even specify an encoding for some Readers/Writers, so whether
something works or not was entirely dependent on default Java
properties.

This modifies the API for XML and JSON Infoset Inputter/Outputters to use
Input/OutputStreams instead of Readers/Writers. This allows the infoset
outputters to control what encoding is used during parsing, and allows
infoset inputters to inspect the preamble to determine the encoding
during unparsing.

The API changes add new new constructors for the Input/OutputStreams and
deprecates the old ones using Reader/Writers.

DAFFODIL-2128