Simple Types no longer have backpointer to elements.
The SimpleType factory classes have been consolidated with the
instance classes, as we no longer need the distinction between them,
though only for simple types so far.

RestrictionFactory and UnionFactory also merged.

Rename parent to lexicalParent in schema compiler.

This is part of changing terminology to make clear that we're
not considering an object referencing a component by QName a "parent".

The term for that is going to be refSources (plural, since there can be
several), and refTarget.

Added toOption to OOLAGValue. Fixes the tests in daffodil-core-unittest
that were broken. Cleans up OOLAG a bit.

Remove hasSep from Element classes. Only sequences can have
separators. They're the only objects we should be asking if they have
separators or not.

Tighten types - DPathElementCompileInfo only points to other of same.

Minor refactor: It is generally better if the classes being inherited
from are abstract. So hoisted the LocalElementDecl into a
LocalElementDeclBase, so that the other flavors of quasi-element aren't
inheriting from a concrete class.

DAFFODIL-1444