Add support for namespacing DFDL extension properties and functions
Create a new dfdlx namespace and move a handful of extensions properties
and functions to it. This is where all future extensions should go.

Extensions that have not been part of an official Daffodil release just
have their namespace changed to dfdlx--backwards compatibility is not
maintained. For extensions that have been part of an official release,
backwards compatibility is maintained, but now results in a deprecation
warning if the old namespace is used. Below is the list of all extension
properties and functions moved to the dfdlx namespace.

Backwards compatibility is supported for:

  dfdl:layerTransform
  dfdl:layerEncoding
  dfdl:layerLengthKind
  dfdl:layerLength
  dfdl:layerLengthUnits
  dfdl:layerBoundaryMark
  daf:parseUnparsePolicy
  daf:trace()
  daf:error()

Backwards compatibility is not supported for:

  dfdl:emptyElementParseUnparsePolicy
  dfdl:inputTypeCalc
  dfdl:outputTypeCalc
  dfdl:repType
  dfdl:repValueRanges
  dfdl:repValues
  dfdl:inputTypeCalcInt()
  dfdl:inputTypeCalcString()
  dfdl:logicalTypeValueInt()
  dfdl:logicalTypeValueString()
  dfdl:outputTypeCalcInt()
  dfdl:outputTypeCalcString()
  dfdl:repTypeValueInt()
  dfdl:repTypeValueString()
  dfdl:outputTypeCalcNextSiblingInt()
  dfdl:outputTypeCalcNextSiblingString()
  daf:lookahead()

The daf namespace does still exist, but is only used for Daffodil
specific things like tunables, configurations files, etc. Only its use
in DFDL schemas is deprecated.

DAFFODIL-2096