Remove the emptyElementParsePolicy warning
This property is a new extension that most schemas will not have. And if
they do not have it, it results in a lot of warnings, seems almost one
for every element. For now, just remove this warning so that users
aren't spammed with it and forced to add an extension property to hide
the warning. If this extension becomes part of the spec, we can reenable
the warning then.

DAFFODIL-2150