Simplify how we show text differences when a TDML test fails
We previously found every character that differed and showed its path,
index, and expected/actual characters. But when an XML element contains
a lot of text (e.g. hexBinary blob) with a lot of differences, this
character by character diff is not very helpful and consumed a lot
memory to build the output string. This easily leads to an OutOfMemory
exception.

Instead, when showing the diff of two XML text elements, only show the
index of the first diff plus some following characters for context. This
should be useful enough in most cases to figure out where things when
off the rails while keep memory usage low, even on large diffs.

Also change the diff output to show the actual infoset before the
differences summary. Displaying the actual infoset after the diff made
it very difficult to find the summary of differences when big infosets
were involved.

DAFFODIL-2118