Warn when DFDL properties are ignored
During schema compilation, a cache of properties exists to speed up
subsequent property lookups. This patch takes advantage of that cache to
determine which properties have been used or not. At the end of schema
compilation, we now inspect all Terms and look for any properties that
are defined on each Term but are not in the cache, which implies that a
property was not used and we can create an SDW. These warnings should
help schema developers realize when properties are accidentally ignored
(e.g. dfdl:lengthPattern is defined but dfdl:lengthKind is not "pattern").

Additionally, lookup functions were refactored to ensure properties are
always cached. Some parts of the code would skip caching if it used
findPropertyOption or other functions instead of cachePropertyOption. To
avoid accidentally not adding a lookup to the cache, the cache*
functions are removed, the findPropertyOption always caches, and all
property lookups are changed to use this function or a wrapper.

Also modify all tests to remove unused properties unless explicitly
testing that a property is ignored.

DAFFODIL-1619