Correct invalid tdml files that should be valid.
Mostly this is missing definitions for ex:, tns: and dfdlx:

This is progress on DAFFODIL-1638.

Removed dfdl-config-format.xsd This was redundantly defining the dafext
namespace. Unusable that way in conjunction with the dafext.xsd file.

Nothing was including/importing this file anyway.

Moved the one definition in it, of dfdlConfig to dafext.xsd.

Fix XML Catalog entries that lacked org/apache/daffodil in paths.

Remove xsi:schemaLocation which breaks CLI tests, due to bug DAFFODIL-2339

DAFFODIL-1638